+++ FE/UI/src/app/components/main_page_component/main-page/Interfaces/Products.ts
@@ -10,4 +10,5 @@ export interface Products {
   name: string;
   category: string;
   price: number;
+  //quantityInStock: number; // add this later
 }
